+++ app/src/main/java/com/example/rpn_calculator/AddToStackCalculatorCommand.kt
@@ -1,6 +1,9 @@
 package com.example.rpn_calculator
 
-class AddToStackCalculatorCommand(private val stack: ArrayDeque<Int>, private val number: Int) :
+class AddToStackCalculatorCommand(
+    private val stack: ArrayDeque<Double>,
+    private val number: Double
+) :
     CalculatorCommand {
 
     override fun execute(): CommandResult {
